% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_concSandFine.R
\name{calc_concSandFine}
\alias{calc_concSandFine}
\title{calc_concSandFine}
\usage{
calc_concSandFine(x, plotTime = FALSE, plotFlow = FALSE)
}
\arguments{
\item{x}{A \code{dataframe} output from \code{get_localNWIS}}

\item{plotTime}{Logical, if \code{TRUE} basic timeseries plots of Conc vs. Time are created for each site with complete data, default is \code{FALSE}}

\item{plotFlow}{Logical, if \code{TRUE} basic plots of Conc vs. Flow are created for each site with complete data, default is \code{FALSE}}
}
\value{
A dataframe containing all samples with complete SSC, sand/fine split, and calculated sand and fine concentrations
}
\description{
Function to calculate sand/fine concentration based on SSC and sand/fine split value and output basic plots
}
\details{
Sand and Fine concentration (mg/L) is calculated for sites with SSC (P80154) and % finer than 62.5 microns (P70331)

Some measure of discharge (P00060, P00061, P30208, P30209, P50042, P72137, P72243, P99060, P99061) must be present a record to plot in plotFlow, otherwise it is removed
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData
concSandFine <- calc_concSandFine(x, plotTime = FALSE, plotFlow = FALSE)

}
