% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_summaryStats.R
\name{calc_summaryStats}
\alias{calc_summaryStats}
\title{calc_summaryStats. Calculate summary stats and tally non-detections and samples with no result values reported.}
\usage{
calc_summaryStats(x, pcodes = c("80154", "70331", "00530", "80155", "80225",
  "91145"))
}
\arguments{
\item{x}{A \code{dataframe} output from \code{get_localNWIS}}

\item{pcodes}{A character vector of parameter codes of interest. 
Default pcodes are SSC (80154), Sand/silt break on suspended (70331), TSS (00530), SSL (80155), Bedload (80225), Bedload mass (91145)}
}
\value{
A dataframe of summary statistics
}
\description{
Calculates summary statistics, and tallys non-detects and NA result values, grouped by site, parameter, and water year.
}
\details{
Calculates number of samples, minimum, maximum, median, mean, and standard deviation (if applicable). 
Non-detects (REMARK_CD = "<") and Averages (REMARK_CD = "A") are not included in calculations. Rejected samples are not included.

A count of the number of non-detect samples and samples where no result value was reported (RESULT_VA = \code{NA})
 is included in the summary. Summary value "n" does not include non-detects or no result reported samples.

Default pcodes are SSC (80154), Sand/silt break on suspended (70331), TSS (00530), SSL (80155), 
Bedload (80225), Bedload mass (91145)

Statistics are separated by Site, parameter, year (WY or user defined), and sample method group.

Sample method groups. Cross section (X-section) is defined as sample method 10(EWI), 15(EWT), or 20(EDI).
Point or non-cross section (Pt/non-X-section) is defined as sample method 30 (single vertical), 40 (multiple verticals), 
50 (point sample), 55 (composite - multiple point samples), 60 (weighted bottle), 70 (grab sample - dip), 100 (Van Dorn), 
900 (SS pumping), 920 (SS BSV DI att), 930 (SS partial depth), 940 (SS partial width), 
4033 (suction lift peristaltic), 4080 (peristaltic pump). Samples with missing or other coded method are grouped together.
}
\examples{
data('exampleData', package = "sedReview")
x <- exampleData
summaryStats <- calc_summaryStats(x)
}
