% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_all.R
\name{check_all}
\alias{check_all}
\title{Function to run all checks on a dataset}
\usage{
check_all(x, returnAllTables = FALSE)
}
\arguments{
\item{x}{A \code{longTable} dataframe output from \code{getLocalNWIS}}

\item{returnAllTables}{Return all tables of flagged results}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to run all checks on a dataset
}
\details{
Runs all check_ functions and outputs a summary dataframe of flagged samples or a list of all flag results if \code{returnAllTables = TRUE}
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData
allChecks <- check_all(x, returnAll = FALSE, returnAllTables = FALSE)
}
