% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bagIE.R
\name{check_bagIE}
\alias{check_bagIE}
\title{check_bagIE. Check and flag bag intake efficiency parameters, and calculate and flag the bag efficiency}
\usage{
check_bagIE(x, returnAll = FALSE, reviewSummary = FALSE)
}
\arguments{
\item{x}{A \code{dataframe} output from \code{get_localNWIS}}

\item{returnAll}{logical, return dataframe containing all results or only return flagged samples. Defualt is FALSE}

\item{reviewSummary}{logical, for center-level review, if \code{TRUE} a summary count of flags by site and water year is returned
instead of individual flagged samples.}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to check and flag bag intake efficiency parameters, and calculate and flag the bag efficiency
}
\details{
checks and calculates bag intake efficiency under OSW Tech Memo 2013.03 \url{https://water.usgs.gov/admin/memo/SW/sw13.03.pdf}

P72217 - Duration sampler collected water, seconds

P72218 - Sample volume to compute isokinetic transit rate, milliliters

P72196 - Velocity to compute isokinetic transit rate, feet per second

P72219 - Sampler nozzle material, code: Plastic=2, TFE=3

P72220 - Sampler nozzle diameter, code: 3/16"=3, 1/4"=4, 5/16"=5

Rejected samples are not included.
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData
bagIEflags <- check_bagIE(x)

}
