% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_commentsNoResult.R
\name{check_commentsNoResult}
\alias{check_commentsNoResult}
\title{check_commentsNoResult. Review unique sample and result comments, and pcodes with no result reported.}
\usage{
check_commentsNoResult(x, pcodes = c("80154", "70331", "00530", "80155",
  "80225", "91145"))
}
\arguments{
\item{x}{A \code{dataframe} output from \code{get_localNWIS}}

\item{pcodes}{A character vector of parameter codes of interest. Default pcodes are SSC (80154), Sand/silt break on suspended (70331), TSS (00530), SSL (80155), Bedload (80225), Bedload mass (91145)}
}
\value{
A dataframe of sample info, comments, and flagged samples with no results reported.
}
\description{
Get unique sample and result comments, and pcodes with no result reported.
}
\details{
Returns unique sample and result comments (\code{SAMPLE_CM_TX} and \code{RESULT_CM_TX} from NWIS) by site and record number.

Returns samples with defined parameter codes where no result was reported (\code{RESULT_VA = NA}) and joins them to comments.
}
\examples{
data('exampleData', package = "sedReview")
checkComments <- check_commentsNoResult(exampleData)
}
