% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_hasQ.R
\name{check_hasQ}
\alias{check_hasQ}
\title{check_hasQ}
\usage{
check_hasQ(x, returnAll = FALSE)
}
\arguments{
\item{x}{A \code{dataframe} output from \code{get_localNWIS}}

\item{returnAll}{Logical. Return dataframe containing all samples if \code{TRUE} or only return samples missing discharge if \code{FALSE}. Default is \code{FALSE}}
}
\value{
A data.frame of samples, what sediment data are available, and then a flag for missing discharge along with what discharge parameter are present
}
\description{
Searches for samples that were collected for sediment data but are missing discharge measurements
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData
checkQOut <- check_hasQ(x,returnAll = FALSE)
}
