% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_metaData.R
\name{check_metaData}
\alias{check_metaData}
\title{check_metaData}
\usage{
check_metaData(x, returnAll = FALSE)
}
\arguments{
\item{x}{A \code{dataframe} output from \code{get_localNWIS}}

\item{returnAll}{logical, return dataframe containing all results or only return flagged samples. Defualt is FALSE}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to perform NWIS 20.xx checks
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData
nwis20flags <- check_metaData(x)

}
