% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_Q.R
\name{check_Q}
\alias{check_Q}
\title{check_Q. Searches for samples that were collected for sediment data but are missing discharge measurements}
\usage{
check_Q(x, includeUV = FALSE, returnAll = FALSE)
}
\arguments{
\item{x}{A \code{dataframe} output from \code{get_localNWIS}}

\item{includeUV}{Logical. If \code{x} was returned from \code{get_UVflow}. 
Run optional flagging for records with P00061 >10\% different from Approved UV flow. Default is \code{FALSE}.}

\item{returnAll}{Logical. Return dataframe containing all samples if \code{TRUE} or only return samples missing discharge if \code{FALSE}. Default is \code{FALSE}}
}
\value{
A data.frame of samples, what sediment data are available, and then a flag for missing discharge along with what discharge parameter are present
}
\description{
Searches for samples that were collected for sediment data but are missing discharge measurements.
Additional options flag Instantaneous Discharge (P00061) in NWIS that are >10% different from Approved Unit Value discharge.
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData
checkQOut <- check_Q(x,includeUV = FALSE, returnAll = FALSE)
}
