% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_qaqcDB.R
\name{check_qaqcDB}
\alias{check_qaqcDB}
\title{check_qaqcDB.
Check if SSC, bedload, or bedload mass are coded as WSQ and in the QA/QC database}
\usage{
check_qaqcDB(x, qa.db = "02", returnAll = FALSE)
}
\arguments{
\item{x}{A \code{dataframe} output from \code{get_localNWIS}}

\item{qa.db}{A character string containing the database number of QA samples}

\item{returnAll}{logical, return dataframe containing all results or only return flagged samples. Defualt is FALSE}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to check if SSC, bedload, or bedload mass are coded as WSQ and in the QA/QC database
}
\details{
Function finds and flags SSC (P80154), bedload (P80225), and bedload mass (P91145) results
that are coded as medium_cd "WSQ" and stored in the user defined QA/QC database (default is qa.db = "02") downloaded with \code{get_localNWIS}

function to test EWI sample collected correctly. Number of verticals between 10 and 20.
}
\examples{
\dontrun{
#NOTE: exampleData contains no QA/QC results and will not return any flagged samples
data("exampleData",package="sedReview")
x <- exampleData
check_qaqcDBOut <- check_qaqcDB(x)
}
}
\seealso{
\code{\link[sedReview]{get_localNWIS}}
}
