% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_samplerType.R
\name{check_samplerType}
\alias{check_samplerType}
\title{check_samplerType. Flag uncommon sampler type}
\usage{
check_samplerType(x, returnAll = FALSE)
}
\arguments{
\item{x}{A \code{dataframe} output from \code{get_localNWIS}}

\item{returnAll}{logical, return dataframe containing all results or only return flagged samples. Default is FALSE}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to flag uncommon sampler type
}
\details{
Function determines most common sampler type from dataset or records and flags all others that are not most common.

Note, there are valid reasons for different sampler types, this is just a check.

Rejected samples are not included.
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData
check_samplerTypeOut <- check_samplerType(x)

}
