% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sedMass.R
\name{check_sedMass}
\alias{check_sedMass}
\title{check_sedMass. Check and flag sediment mass (P91157) less than 2mg}
\usage{
check_sedMass(x, returnAll = FALSE, reviewSummary = FALSE)
}
\arguments{
\item{x}{A \code{dataframe} output from \code{get_localNWIS}}

\item{returnAll}{logical, return dataframe containing all results or only return flagged samples. Default is FALSE}

\item{reviewSummary}{logical, for center-level review, if \code{TRUE} a summary count of flags by site and water year is returned
instead of individual flagged samples.}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to check and flag sediment mass (P91157) less than 2mg
}
\details{
Checks parameter code 91157 (Suspended sediment, sediment mass recovered from whole water sample, dry weight, grams)
and flags samples < 2mg (< 0.002 g) with flag and result value.

Rejected samples are not included.
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData
check_sedMassOut <- check_sedMass(x)

}
