% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tss.R
\name{check_tss}
\alias{check_tss}
\title{check_tss. Flag TSS results without an accompanying SSC result}
\usage{
check_tss(x, returnAll = FALSE)
}
\arguments{
\item{x}{A \code{dataframe} output from \code{get_localNWIS}}

\item{returnAll}{logical, return dataframe containing all results or only return flagged samples. Defualt is FALSE}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to flag TSS results without an accompanying SSC result
}
\details{
See OSW Technical Memo No. 2001.03 for more details and background information
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData
check_tssOut <- check_tss(x)

}
