% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_verticals.R
\name{check_verticals}
\alias{check_verticals}
\title{check_verticals. Check if number of verticals in EDI/EWI samples is correct}
\usage{
check_verticals(x, returnAll = FALSE, reviewSummary = FALSE)
}
\arguments{
\item{x}{A \code{dataframe} output from \code{get_localNWIS}}

\item{returnAll}{logical, return dataframe containing all results or only return flagged samples. Default is FALSE}

\item{reviewSummary}{logical, for center-level review, if \code{TRUE} a summary count of flags by site and water year is returned
instead of individual flagged samples.}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to check if number of verticals in EDI/EWI samples is correct
}
\details{
function to test EDI sample collected correctly. Number of sampling verticals 4-9.

function to test EWI sample collected correctly. Number of verticals between 10 and 20.

Rejected samples are not included.
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData
check_verticalsOut <- check_verticals(x)

}
