% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCalcBagIE.R
\name{checkCalcBagIE}
\alias{checkCalcBagIE}
\title{checkCalcBagIE}
\usage{
checkCalcBagIE(x, returnAll = FALSE)
}
\arguments{
\item{x}{A \code{longTable} dataframe output from \code{getLocalNWIS}}

\item{returnAll}{logical, return dataframe containing all results or only return flagged samples. Defualt is FALSE}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to check and flag bag intake efficiency parameters, and calculate and flag the bag efficiency
}
\details{
checks and calculated bag intake efficiency under OSW Tech Memo 2013.03

P72217 - Duration sampler collected water, seconds

P72218 - Sample volume to compute isokinetic transit rate, milliliters

P72196 - Velocity to compute isokinetic transit rate, feet per second

P72219 - Sampler nozzle material, code: Plastic=2, TFE=3

P72220 - Sampler nozzle diameter, code: 3/16"=3, 1/4"=4, 5/16"=5
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData$longTable
bagIEflags <- checkCalcBagIE(x)

}
