% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNWIS20.R
\name{checkNWIS20}
\alias{checkNWIS20}
\title{checkNWIS20}
\usage{
checkNWIS20(x, returnAll = FALSE)
}
\arguments{
\item{x}{A \code{longTable} dataframe output from \code{getLocalNWIS}}

\item{returnAll}{logical, return dataframe containing all results or only return flagged samples. Defualt is FALSE}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to perform NWIS 20.xx checks
}
\examples{
data("testData",package="sedReview")
x <- testData
nwis20flags <- checkNWIS20(x)

}
