% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkOSW2001_03.R
\name{checkOSW2001_03}
\alias{checkOSW2001_03}
\title{checkOSW2001_03}
\usage{
checkOSW2001_03(x, returnAll = FALSE)
}
\arguments{
\item{x}{A \code{longTable} dataframe output from \code{getLocalNWIS}}

\item{returnAll}{logical, return dataframe containing all results or only return flagged samples. Defualt is FALSE}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to flag TSS results without an accompanying SSC result
}
\details{
See OSW Technical Memo No. 2001.03 for more details and background information
}
\examples{
data("testData",package="sedReview")
x <- testData
checkOSW2001_03Flags <- checkOSW2001_03(x)

}
