% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkQ.R
\name{checkQ}
\alias{checkQ}
\title{checkQ}
\usage{
checkQ(x, returnAll = FALSE)
}
\arguments{
\item{x}{A \code{longTable} dataframe output from \code{getLocalNWIS}}

\item{returnAll}{Logical. Return dataframe containing all samples if \code{TRUE} or only return samples missing discharge if \code{FALSE}. Defualt is \code{FALSE}}
}
\value{
A data.frame of samples, what sediment data are available, and then a flag for missing discharge along with what discharge parameter are present
}
\description{
Searches for samples that were collected for sediment data but are missing discharge measurements
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData$longTable
checkQ(x,returnAll = FALSE)
}
