% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSamplerTyp.R
\name{checkSamplerTyp}
\alias{checkSamplerTyp}
\title{checkSamplerTyp}
\usage{
checkSamplerTyp(x, returnAll = FALSE)
}
\arguments{
\item{x}{A \code{longTable} dataframe output from \code{getLocalNWIS}}

\item{returnAll}{logical, return dataframe containing all results or only return flagged samples. Defualt is FALSE}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to flag uncommon sampler type
}
\details{
Function determines most common sampler type from dataset or records and flags all others that are not most common.

Note, there are valid reasons for different sampler types, this is just a check.
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData
samplerTypFlags <- checkSamplerTyp(x)

}
