% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSamPurp.R
\name{checkSamPurp}
\alias{checkSamPurp}
\title{checkSamPurp}
\usage{
checkSamPurp(x, returnAll = FALSE)
}
\arguments{
\item{x}{A \code{longTable} dataframe output from \code{getLocalNWIS}}

\item{returnAll}{logical, return dataframe containing all results or only return flagged samples. Defualt is FALSE}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to flag uncommon sample purpose
}
\details{
Function determines most common sample purpose for all retrieved records and flags all sample purposes other than the most common.

There are of course valid reasons to have different sample purposes, this is just a check.
}
\examples{
data("testData",package="sedReview")
x <- testData
samplePurpFlags <- checkSamPurp(x)

}
