% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_activeSed.R
\name{count_activeSed}
\alias{count_activeSed}
\title{For higher level review. Retrieves all stream site sediment samples in a given date range via an ODBC connection to an internal NWIS server,
and returns a summary table by site, water year, and number of samples in each analysis.}
\usage{
count_activeSed(DSN, env.db = "01", begin.date = NA, end.date = NA)
}
\arguments{
\item{DSN}{A character string containing the DSN for your local server}

\item{env.db}{A character string containing the database number of environmental samples}

\item{begin.date}{Character string containing beginning date of data pull (yyyy-mm-dd)}

\item{end.date}{Character string containing ending date of data pull (yyyy-mm-dd)}
}
\value{
A summary dataframe
}
\description{
For higher level review. Retrieves all stream site sediment samples in a given date range via an ODBC connection to an internal NWIS server,
and returns a summary table by site, water year, and number of samples in each analysis.
}
\details{
For review at the center level. Function retrieves all stream site (SITE_TYP_CD = 'ST') sediment 
(SSC (80154), sand/silt break (70331), TSS (00530), and bedload (80225)) samples in the defined date range of the users internal NWIS database.
A summary table of sample counts by site, water year, and parameter is returned.
}
\examples{
\dontrun{
#Will not run unless connected to NWISCO
activeSed <- count_activeSed(DSN = 'nwisco',
                             env.db = '01',
                             begin.date = '2015-10-01',
                             end.date = '2017-09-30')}
                             
}
