% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_methodsBySite.R
\name{count_methodsBySite}
\alias{count_methodsBySite}
\title{count_methodsBySite}
\usage{
count_methodsBySite(x)
}
\arguments{
\item{x}{A \code{longTable} dataframe output from \code{getLocalNWIS}}
}
\value{
A data.frame tabular summary of counts of samples by site and method
}
\description{
Returns counts of samples by site, sampling method, and sampler type. See details
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData
count_methodsBySiteOut <- count_methodsBySite(x)
}
