% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countStatus.R, R/countbySite.R
\name{countMethodsBySite}
\alias{countMethodsBySite}
\alias{countMethodsBySite}
\title{countStatus}
\usage{
countMethodsBySite(x)

countMethodsBySite(x)
}
\arguments{
\item{x}{A \code{longTable} dataframe output from \code{getLocalNWIS}}

\item{bySite}{Logical. Count by site if \code{TRUE} or for all sites if \code{FALSE}}

\item{x}{A \code{longTable} dataframe output from \code{getLocalNWIS}}
}
\value{
A data.frame tabular summary of counts of sample status (DQI_CD) by parameter

A data.frame tabular summary of counts of samples by site and method
}
\description{
Returns counts of sample status (DQI_CD) by parameter for each site or for multiple sites together

Returns counts of samples by site, sampling method, and sampler type. See details
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData$longTable
countStatus(x,bySite = TRUE)
data("exampleData",package="sedReview")
x <- exampleData$longTable
countMethodsBySite(x)
}
