% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countSamplesByMethod.R
\name{countSamplesByMethod}
\alias{countSamplesByMethod}
\title{countSamplesByMethod}
\usage{
countSamplesByMethod(x)
}
\arguments{
\item{x}{A list generated from \code{getLocalNWIS}}
}
\value{
A data.frame tabular summary of counts of samples by site
}
\description{
Function to provide counts of samples collected using various methods for 
sample method codes 82398 and 82398 and for sampler type code 84164
}
\examples{
x <- data("exampleData",package="sedReview")
countSamplesByMethod(x)
}
