% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countSamplesbySite.R
\name{countSamplesbySite}
\alias{countSamplesbySite}
\title{countSamplesbySite}
\usage{
countSamplesbySite(x)
}
\arguments{
\item{x}{A list generated from \code{getLocalNWIS}}
}
\value{
A data.frame tabular summary of counts of samples by site and method
}
\description{
Returns counts of samples by site, sampling method, and sampler type. See details
}
\examples{
x <- data("exampleData",package="sedReview")
countSamplesbySite(x)
}
