% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countStatus.R
\name{countStatus}
\alias{countStatus}
\title{countStatus}
\usage{
countStatus(x, bySite = TRUE)
}
\arguments{
\item{x}{A \code{longTable} dataframe output from \code{getLocalNWIS}}

\item{bySite}{Logical. Count by site if \code{TRUE} or for all sites if \code{FALSE}}
}
\value{
A data.frame tabular summary of counts of sample status (DQI_CD) by parameter
}
\description{
Returns counts of sample status (DQI_CD) by parameter for each site or for multiple sites together
}
