% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outliers.R
\name{find_outliers}
\alias{find_outliers}
\title{find_outliers}
\usage{
find_outliers(x, lowThreshold = 0.1, highThreshold = 0.9,
  returnAll = FALSE)
}
\arguments{
\item{x}{A \code{longTable} dataframe output from \code{getLocalNWIS}}

\item{lowThreshold}{numeric value between 0 and 1 indicating the quantile threshold for a low value outlier.}

\item{highThreshold}{numeric value between 0 and 1 indicating the quantile threshold for a high value outlier.}

\item{returnAll}{logical, return dataframe containing all results or only return flagged samples. Defualt is FALSE}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to flag sediment outliers
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData
find_outliersOut <- find_outliers(x)

}
