% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_provisional.R
\name{find_provisional}
\alias{find_provisional}
\title{find_provisional}
\usage{
find_provisional(x, env.db = "01", qa.db = "02", pcodes = c("80154",
  "70331", "00530", "80155", "80225", "91145"), env.fileout = NULL,
  qa.fileout = NULL)
}
\arguments{
\item{x}{A \code{dataframe} output from \code{get_localNWIS}}

\item{env.db}{A character string containing the database number of environmental samples}

\item{qa.db}{A character string containing the database number of QA samples}

\item{pcodes}{A character vector of parameter codes of interest. Default pcodes are SSC (80154), Sand/silt break on suspended (70331), TSS (00530), SSL (80155), Bedload (80225), Bedload mass (91145)}

\item{env.fileout}{A character string of the filename for environmental database records to be output. Default is NULL, if NULL a file will not be output.}

\item{qa.fileout}{A character string of the filename for QA database records to be output. Default is NULL, if NULL a file will not be output.}
}
\value{
A list of dataframes for the environmental and QA database records. List names are env.provisional and qa.provisional
}
\description{
Find sediment parameter code records with DQI code of "S" and output record number files for QWDATA.
}
\details{
In RStudio the function will automatically show the environmental and QA database records in view tabs.

Default pcodes are SSC (80154), Sand/silt break on suspended (70331), TSS (00530), SSL (80155), Bedload (80225), Bedload mass (91145)
}
\examples{
data('exampleData', package = "sedReview")
x <- exampleData
provisional.recs <- find_provisional(x)
\dontrun{
view(provisional.recs$env.provisional)}
}
