% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagSummary.R
\name{flagSummary}
\alias{flagSummary}
\title{Function to summarize samples that are not approved}
\usage{
flagSummary(x, returnAll = FALSE, returnAllTables = FALSE)
}
\arguments{
\item{x}{A \code{longTable} dataframe output from \code{getLocalNWIS}}

\item{returnAll}{Only return results that have been flagged somewhere}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to summarize samples that are not approved
}
\details{
Summarizes automated check tables and displays records numbers for unapproved data (DQI of I, S, or P) and number of flags
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData
flagSummaryOut <- flagSummary(x, returnAll = FALSE)
}
