% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_UVflow.R
\name{get_UVflow}
\alias{get_UVflow}
\title{get_UVflow. Joins UV flow data from NWISweb to SED data from local NWIS database.}
\usage{
get_UVflow(x, max.diff = "1 hours")
}
\arguments{
\item{x}{A \code{dataframe} output from \code{get_localNWIS}}

\item{max.diff}{the maximum allowable difference in time for a match. See \bold{Details}.
Default is "1 hour" (ie. look for a paired UV 1 hour before and 1 hour after a SED sample timestamp)}
}
\value{
A dataframe containing all data originally in dataframe \code{x},
with available Approved Unit Value discharge appended from NWISweb.
}
\description{
Joins UV flow data from NWISweb to SED data from local NWIS database. Note this function can take a while to run.
}
\details{
This function utilizes the \code{dataRetrieval::readSWISuv} to pull Approved unit value discharge values from NWISweb
and join them with SED records from the local NWIS data pull based on sampling time and unit value time.

Options utlizing the UV flow are contained in the \code{check_Q} and \code{find_boxcoef} functions.

NOTE: the \code{dataRetrieval::readWISuv} utilizes all sites and the maximum date range in 
dataframe \code{x} from \code{get_localNWIS}.
Large numbers of sites or long date ranges can take several minutes to complete.

The format for \code{max.diff} should be a numeric value followed by a description of the time span.
The time span must be one of "secs", "mins", "hours", "days", or "weeks" for seconds, minutes, hours, days, or weeks, respectively.

The merge of data by time borrows code from \code{smwr::Base}, but the package is not a dependency.
}
\examples{
\dontrun{
data("exampleData",package="sedReview")
x <- exampleData
mergeUV <- get_UVflow(x, max.diff = "1 hours")
}
}
\seealso{
\code{\link[sedReview]{get_localNWIS}}, \code{\link[dataRetrieval]{readNWISuv}}
}
