% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_wideTable.R
\name{make_wideTable}
\alias{make_wideTable}
\title{make_wideTable}
\usage{
make_wideTable(x)
}
\arguments{
\item{x}{A \code{longTable} dataframe output from \code{getLocalNWIS}}
}
\value{
A wide format data.frame
}
\description{
Reformats teh longTable output (by result) from getLocalNWIS into wide format (by sample)
}
\details{
The wide format data frame format is primarily for visual inspection of data.
A number of important metadata are missing from this table and thus it is not recommended to use these data in analyis
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData
make_wideTableOut <- make_wideTable(x)
}
