% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sedFlow.R
\name{plot_sedFlow}
\alias{plot_sedFlow}
\title{plot_sedFlow. Create plots of sediment vs. discharge for select sediment parameters.}
\usage{
plot_sedFlow(x, log.P80154 = FALSE, log.P70331 = FALSE,
  log.P80225 = FALSE, log.P00530 = FALSE, log.Q = FALSE,
  siteSelect = NULL, PDFout = NULL)
}
\arguments{
\item{x}{A \code{dataframe} output from \code{get_localNWIS}}

\item{log.P80154}{Logical, if \code{TRUE}, SSC y axis will be log10. Default is \code{FALSE}.}

\item{log.P70331}{Logical, if \code{TRUE}, sand/silt break y axis will be log10. Default is \code{FALSE}.}

\item{log.P80225}{Logical, if \code{TRUE}, bedload y axis will be log10. Default is \code{FALSE}.}

\item{log.P00530}{Logical, if \code{TRUE}, TSS y axis will be log10. Default is \code{FALSE}}

\item{log.Q}{Logical, if \code{TRUE}, flow x axis will be log10 for all parameters present. Default is \code{FALSE}.}

\item{siteSelect}{Character, site number to create plots for if \code{x} contains multiple sites. Default is \code{NULL}.}

\item{PDFout}{Character. File or full path name of file for plots. If \code{NULL}, the default, a list of the plots will be returned in R instead.}
}
\value{
If \code{PDFout = NULL}, list containing ggplot elements. If \code{PDFout} specified, a PDF document containing the plots.
}
\description{
Function to output scatterplots of sediment parameter vs. discharge. Output is list of plots or write to PDF.
}
\details{
Scatterplots of SSC (P80154), sand/silt break (P70331), bedload (P80225), TSS (P00530) vs. flow (all discharge reported or converted to cfs).
Not all parameters need to be present in \code{x} for plots to be created (eg. if only SSC present than only SSC vs. flow plot returned)

If PDFout is not specified, than a list of the plots is returned. Plots (if applicable) are specified above. See example for more details.
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData
sedFlow <- plot_sedFlow(exampleData, siteSelect = "06934500")
\dontrun{
# view plot in Rstudio
sedFlow$SSC

# output to file on D drive
plot_sedFlow(exampleData, siteSelect = "06934500", PDFout = "D:/flow.pdf")
}

}
