% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SSCTSS.R
\name{plot_SSCTSS}
\alias{plot_SSCTSS}
\title{plot_SSCTSS. Plot individual and side-by-side boxplots for SSC and TSS, and SSC vs. TSS scatterplot at a site.}
\usage{
plot_SSCTSS(x, siteSelect = NULL, PDFout = NULL)
}
\arguments{
\item{x}{A \code{dataframe} output from \code{get_localNWIS}}

\item{siteSelect}{Character, site number to create plots for if \code{x} contains multiple sites. Default is \code{NULL}.}

\item{PDFout}{Character. File or full path name of file for plots. If \code{NULL}, the default, a list of the plots will be returned in R instead.}
}
\value{
If \code{PDFout = NULL}, list containing ggplot elements. If \code{PDFout} specified, a PDF document containing the plots.
}
\description{
Function to output individual and side-by-side boxplots, and scatterplot, for SSC and TSS at a site. Output is list of plots or write to PDF.
}
\details{
Boxplots of SSC (P80154) and TSS (P00530). Box contains lower, median, and upper quartile. Whiskers extend to farthest point
within +/- 1.5 IQR. Any points outside of +/- 1.5 IQR are plotted as outlier points.

If PDFout is not specified, than a list of the plots is returned. Plots (if applicable) are boxplot of SSC, boxplot of TSS,
side-by-side boxplot of SSC and TSS, and a SSC vs. TSS scatterplot of samples with both analyses. See example for more details.

Portions of code modified from \code{WQReview::qwparmBoxPlot}.
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData2
ssctss <- plot_SSCTSS(exampleData2, siteSelect = "09163500")
\dontrun{
# view plot in Rstudio
ssctss$combined

# output to file on D drive
plot_ssctss(exampleData2, siteSelect = "09163500", PDFout = "D:/ssctss.pdf")
}

}
