% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_boxcoef.R
\name{summary_boxcoef}
\alias{summary_boxcoef}
\title{summary_boxcoef. Summarise box coefficient pairs for multiple sites and years utilizing the find_boxcoef function}
\usage{
summary_boxcoef(x, timediff = 1, methods_NX = c(30, 40, 50, 55, 70, 100,
  900, 920, 930, 940, 4033, 4080), methods_X = c(10, 15, 20),
  includeUV = FALSE, returnAllTables = FALSE)
}
\arguments{
\item{x}{A \code{dataframe} output from \code{get_localNWIS}}

\item{timediff}{Number of hours to look before and after a point sample for a paired cross-section sample.
Default is 1 (ie. look for a paired samples 1 hour before and 1 hour after a non-cross section/point sample timestamp)}

\item{methods_NX}{Vector of Sampling Method (P82398) values that define non-cross section/point samples of the box coefficient.
Default is:\cr\cr methods_NX = c(30,40,50,55,70,100,900,920,930,940,4033,4080)\cr\cr 
30 (single vertical), 40 (multiple verticals), 50 (point sample), 55 (composite - multiple point samples), 
60 (weighted bottle), 70 (grab sample - dip), 100 (Van Dorn), 900 (SS pumping), 
920 (SS BSV DI att), 930 (SS partial depth), 940 (SS partial width), 4033 (suction lift peristaltic), 4080 (peristaltic pump).}

\item{methods_X}{Vector of Sampling Method (P82398) values that define cross section samples of the box coefficient.
Default is:\cr\cr methods_X = c(10,15,20)\cr\cr
10 (EWI), 15 (multiple verticals non-isokinetic EWT), or 20 (EDI).}

\item{includeUV}{Logical. If \code{x} was returned from \code{get_UVflow}, 
run optional addition of available Approved Unit Value discharge to output table. Default is \code{FALSE}.}

\item{returnAllTables}{Logical. If \code{TRUE}, return a list with the summary table and all individual site tables from the find_boxcoef function used internally.
Default is FALSE.}
}
\value{
A dataframe of the summary counts of box coefficient pairs, or a list of the summary and individual site box_coef data
}
\description{
Function to summarise box coefficient pairs for multiple sites and years utilizing the find_boxcoef function
}
\details{
Function summarises box coefficient pairs by each site and WY. See documentation for \code{find_boxcoef} function for more details.
}
\examples{
data('exampleData2', package = "sedReview")
x <- exampleData2
boxCoef <- summary_boxcoef(x)
boxCoef_list <- summary_boxcoef(x, returnAllTables = TRUE)
\dontrun{
View(boxCoef_list$summary)
}
}
\seealso{
\code{\link[sedReview]{get_localNWIS}}, \code{\link[sedReview]{find_boxcoef}}, \code{\link[sedReview]{get_UVflow}}
}
