% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_concSandFine.R
\name{calc_concSandFine}
\alias{calc_concSandFine}
\title{calc_concSandFine. calculate sand/fine concentration based on SSC and sand/fine split value}
\usage{
calc_concSandFine(x)
}
\arguments{
\item{x}{A \code{dataframe} output from \code{get_localNWIS}}
}
\value{
A dataframe containing all samples with complete SSC, sand/fine split, and calculated sand and fine concentrations
}
\description{
Function to calculate sand/fine concentration based on SSC and sand/fine split value
}
\details{
Sand and Fine concentration (mg/L) is calculated for sites with SSC (P80154) and % finer than 62.5 microns (P70331)

Rejected samples are not included.
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData
concSandFine <- calc_concSandFine(x)

}
