% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_summaryStats.R
\name{calc_summaryStats}
\alias{calc_summaryStats}
\title{calc_summaryStats. Calculate summary stats and non-detections.}
\usage{
calc_summaryStats(x, pcodes = c("80154", "70331", "00530", "80155", "80225",
  "91145"))
}
\arguments{
\item{x}{A \code{dataframe} output from \code{get_localNWIS}}

\item{pcodes}{A character vector of parameter codes of interest. Default pcodes are SSC (80154), Sand/silt break on suspended (70331), TSS (00530), SSL (80155), Bedload (80225), Bedload mass (91145)}
}
\value{
A dataframe of summary statistics
}
\description{
Calculates summary statistics and tallys non-detects grouped by site, parameter, and water year.
}
\details{
Calculates number of samples, minimum, maximum, median, mean, and standard deviation (if applicable). Non-detects (REMARK_CD = "<") and Averages (REMARK_CD = "A") are not included in calculations.

A count of the number of non-detect samples is included in the summary. Summary value "n" does not include non-detects.

Default pcodes are SSC (80154), Sand/silt break on suspended (70331), TSS (00530), SSL (80155), Bedload (80225), Bedload mass (91145)
}
\examples{
data('exampleData', package = "sedReview")
x <- exampleData
summaryStats <- calc_summaryStats(x)
}
