% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_all.R
\name{check_all}
\alias{check_all}
\title{check_all. Run all project level review checks, counts, finds, and calcs on a dataset.}
\usage{
check_all(x, qa.db = "02", includeUV = FALSE, returnAllTables = FALSE)
}
\arguments{
\item{x}{A \code{dataframe} output from \code{get_localNWIS}}

\item{qa.db}{A character string containing the database number of QA samples (for check_qaqcDB function), should be same QA DB used in \code{get_localNWIS}}

\item{includeUV}{Logical. If \code{x} was returned from \code{get_UVflow}. 
Run optional flagging in \code{check_Q}.}

\item{returnAllTables}{Return all tables of flagged results}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to run all project level review checks, counts, finds, and calcs on a dataset.
}
\details{
Runs all check_, count_, and find_ functions and outputs a summary dataframe of flagged samples or a list of all flag results if \code{returnAllTables = TRUE}
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData
allChecks <- check_all(x, returnAllTables = FALSE)
allChecksList <- check_all(x, returnAllTables = TRUE)
\dontrun{
view(allChecksList$flaggedSamples)
}
}
\seealso{
\code{\link[sedReview]{check_bagIE}},\code{\link[sedReview]{check_commentsNoResult}}, 
\code{\link[sedReview]{check_Q}}, \code{\link[sedReview]{check_metaData}},
\code{\link[sedReview]{check_samplePurp}}, \code{\link[sedReview]{check_samplerType}}, \code{\link[sedReview]{check_sedMass}}, 
\code{\link[sedReview]{check_tss}}, \code{\link[sedReview]{check_verticals}}, \code{\link[sedReview]{check_qaqcDB}}
\code{\link[sedReview]{count_methodsBySite}}, \code{\link[sedReview]{count_sampleStatus}},
\code{\link[sedReview]{find_boxcoef}} \code{\link[sedReview]{find_outliers}},
\code{\link[sedReview]{find_provisional}}, \code{\link[sedReview]{calc_concSandFine}}, \code{\link[sedReview]{calc_summaryStats}}
}
