% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_all.R
\name{check_all}
\alias{check_all}
\title{check_all}
\usage{
check_all(x, returnAllTables = FALSE)
}
\arguments{
\item{x}{A \code{dataframe} output from \code{get_localNWIS}}

\item{returnAllTables}{Return all tables of flagged results}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to run all checks on a dataset
}
\details{
Runs all check_, count_, and find_ functions and outputs a summary dataframe of flagged samples or a list of all flag results if \code{returnAllTables = TRUE}
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData
allChecks <- check_all(x, returnAllTables = FALSE)
allChecksList <- check_all(x, returnAllTables = TRUE)
view(allChecksList$flaggedSamples)
}
\seealso{
\code{\link[sedReview]{check_bagIE}}, \code{\link[sedReview]{check_hasQ}}, \code{\link[sedReview]{check_metaData}},
\code{\link[sedReview]{check_samplePurp}}, \code{\link[sedReview]{check_samplerType}}, 
\code{\link[sedReview]{check_tss}}, \code{\link[sedReview]{check_verticals}}, \code{\link[sedReview]{check_qaqcDB}}
\code{\link[sedReview]{count_methodsBySite}}, \code{\link[sedReview]{count_sampleStatus}}, \code{\link[sedReview]{find_outliers}}
}
