% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_verticals.R
\name{check_verticals}
\alias{check_verticals}
\title{check_verticals}
\usage{
check_verticals(x, returnAll = FALSE)
}
\arguments{
\item{x}{A \code{longTable} dataframe output from \code{getLocalNWIS}}

\item{returnAll}{logical, return dataframe containing all results or only return flagged samples. Defualt is FALSE}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to check if number of verticals in EDI/EWI samples is correct
}
\details{
function to test EDI sample collected correctly. Number of sampling verticals 4-9.

function to test EWI sample collected correctly. Number of verticals between 10 and 20.
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData
check_verticalsOut <- check_verticals(x)

}
