% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNWIS20.R
\name{checkNWIS20}
\alias{checkNWIS20}
\title{checkNWIS20}
\usage{
checkNWIS20(x, returnAll = FALSE)
}
\arguments{
\item{returnAll}{logical, return dataframe containing all results or only return flagged samples. Defualt is FALSE}

\item{qw.data}{A qw.data list generated from readNWISodbc}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to perform NWIS 20.xx checks
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData$PlotTable
nwis20flags <- checkNWIS20(x)

}
