% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSamplerTyp.R
\name{checkSamplerTyp}
\alias{checkSamplerTyp}
\title{checkSamplerTyp}
\usage{
checkSamplerTyp(x, returnAll = FALSE)
}
\arguments{
\item{x}{A \code{longTable} dataframe output from \code{getLocalNWIS}}

\item{returnAll}{logical, return dataframe containing all results or only return flagged samples. Defualt is FALSE}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to flag uncommon sampler type
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData$longTable
samplerTypFlags <- checkSamplerTyp(x)

}
