% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSamPurp.R
\name{checkSamPurp}
\alias{checkSamPurp}
\title{checkSamPurp}
\usage{
checkSamPurp(x, returnAll = FALSE)
}
\arguments{
\item{returnAll}{logical, return dataframe containing all results or only return flagged samples. Defualt is FALSE}

\item{qw.data}{A qw.data list generated from readNWISodbc}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to flag uncommon sample purpose
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData$PlotTable
samplePurpFlags <- checkSamPurp(x)

}
