% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_methodsBySite.R
\name{count_methodsBySite}
\alias{count_methodsBySite}
\title{count_methodsBySite. Counts of samples by site, WY, sampling method, and sampler type.}
\usage{
count_methodsBySite(x)
}
\arguments{
\item{x}{A \code{dataframe} output from \code{get_localNWIS}}
}
\value{
A data.frame tabular summary of counts of samples by site, water year, and method
}
\description{
Returns counts of samples by site, WY, sampling method, and sampler type.

Returns total counts of SSC(P80154), Bedload(P80225), Bedload mass (P91145), and the following specifics:

counts of SSC methode code (P82398) = 10, 20, or 60

counts of SSC sampler type (P84164) = 3060, 3070, or 3071

counts of Bedload method code (P82398) = 1000, 1010, or 1020
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData
count_methodsBySiteOut <- count_methodsBySite(x)
}
