% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_methodsBySite.R
\name{count_methodsBySite}
\alias{count_methodsBySite}
\title{count_methodsBySite. Counts of samples by site, WY, sampling method, and sampler type.}
\usage{
count_methodsBySite(x)
}
\arguments{
\item{x}{A \code{dataframe} output from \code{get_localNWIS}}
}
\value{
A data.frame tabular summary of counts of samples by site, water year, and method
}
\description{
Returns counts of samples by site, WY, sampling method, and sampler type.

Returns total counts of SSC(P80154), Bedload(P80225), Bedload mass (P91145), and the following specifics:

counts of SSC methode code (P82398)

counts of SSC sampler type (P84164)

counts of Bedload method code (P82398)

Note: Counts include rejected and no-results-reported samples. See \code{check_metaData}
for samples missing method or sampler coding.
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData
count_methodsBySiteOut <- count_methodsBySite(x)
}
\seealso{
\code{\link[sedReview]{check_metaData}}
}
