% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMethodsBySite.R
\name{countMethodsBySite}
\alias{countMethodsBySite}
\title{countMethodsBySite}
\usage{
countMethodsBySite(x)
}
\arguments{
\item{x}{A \code{longTable} dataframe output from \code{getLocalNWIS}}
}
\value{
A data.frame tabular summary of counts of samples by site and method
}
\description{
Returns counts of samples by site, sampling method, and sampler type. See details
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData$longTable
countMethodsBySite(x)
}
