% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outliers.R
\name{find_outliers}
\alias{find_outliers}
\title{find_outliers. Flag sediment outliers based on threshold criteria.}
\usage{
find_outliers(x, site_no = NULL, lowThreshold = 0.1, highThreshold = 0.9,
  returnAll = FALSE)
}
\arguments{
\item{x}{A \code{dataframe} output from \code{get_localNWIS}}

\item{site_no}{Character of a site number in the x \code{dataframe} from \code{get_localNWIS} if x contains more than one site. Default is NULL.}

\item{lowThreshold}{numeric value between 0 and 1 indicating the quantile threshold for a low value outlier.}

\item{highThreshold}{numeric value between 0 and 1 indicating the quantile threshold for a high value outlier.}

\item{returnAll}{logical, return dataframe containing all results or only return flagged samples. Defualt is FALSE}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to flag sediment outliers based on threshold criteria.

Flags for suspended sediment concentration (P80154), suspended sediment load (P80155), bed and suspended sediment loss on ignition (P00496/00535), and Sand/Silt break (P70331 Suspended sediment, sieve diamter, percent smaller than 0.0625mm)
}
\details{
Rejected samples are not included.
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData
outliers_05586300 <- find_outliers(x, site_no = "05586300")

}
