% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_localNWIS.R
\name{get_localNWIS}
\alias{get_localNWIS}
\title{get_localNWIS.  
Pulls data from NWIS internal servers using ODBC connection and returns a longtable (by result)}
\usage{
get_localNWIS(DSN, env.db = "01", qa.db = "02", STAIDS, begin.date = NA,
  end.date = NA, projectCd = NULL, resultAsText = FALSE,
  approval = "All")
}
\arguments{
\item{DSN}{A character string containing the DSN for your local server}

\item{env.db}{A character string containing the database number of environmental samples}

\item{qa.db}{A character string containing the database number of QA samples}

\item{STAIDS}{A character vector of stations IDs. Agency code defaults to "USGS" unless appended to the beginning of station ID with a dash, e.g. "USGS-123456".}

\item{begin.date}{Character string containing beginning date of data pull (yyyy-mm-dd)}

\item{end.date}{Character string containing ending date of data pull (yyyy-mm-dd)}

\item{projectCd}{Character vector containing project codes to subset data by.}

\item{resultAsText}{Output results as character instead of numeric. Used for literal output of results from NWIS that are no affected by class coerrcion, such as dropping zeros after decimal point. Default is False.}

\item{approval}{A character string indicating which DQI samples to pull.\cr
Default is 'All', which includes approved, rejected, and in-review samples.\cr Other options are:\cr
"Rejected" which pulls only samples with DQI = "Q","X", or "U".\cr 
"Non-rejected" which includes any historical, approved, provisional, and in-review samples.}
}
\value{
Returns a dataframe of samples. 
The longTable format contains all data pulled from NWIS along with all assosciated metadata in by-result format. 
For a wideTable containing all data pulled from NWIS in wide (sample-result) format, an easier format for import into spreadsheet programs, use the \code{make_wideTable} function
Dataframe names will be changed to more appropriate values in future package updates.
}
\description{
get_localNWIS.  
Pulls data from NWIS internal servers using ODBC connection and returns a longtable (by result)
}
\details{
NWIS parameter groups are as follows: All = "All",
physical = "PHY",
cations = "INM",
anions = "INN",
nutrients = "NUT",
microbiological = "MBI",
biological = "BIO",
metals = "IMM",
nonmetals = "IMN",
pesticides = "TOX",
pcbs="OPE",
other organics = "OPC",
radio chemicals = "RAD",
stable isotopes = "XXX",
sediment = "SED",
population/community = "POP"
}
\examples{
\dontrun{
#Will not run unless connected to NWISCO
qw.data <- get_localNWIS(DSN="NWISCO",
                             env.db = "01",
                             qa.db = "02",
                             STAIDS = c("06733000","09067005"),
                             begin.date = "2005-01-01",
                             end.date = "2015-10-27",
                             projectCd = NULL,
                             resultAsText = FALSE,
                             approval = 'All')
}
}
