% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSedOutliers.R
\name{getSedOutliers}
\alias{getSedOutliers}
\title{getSedOUtliers}
\usage{
getSedOutliers(x, lowThreshold = 0.1, highThreshold = 0.9,
  returnAll = FALSE)
}
\arguments{
\item{x}{A \code{longTable} dataframe output from \code{getLocalNWIS}}

\item{returnAll}{logical, return dataframe containing all results or only return flagged samples. Defualt is FALSE}
}
\value{
A dataframe containing all samples with applicable flags
}
\description{
Function to flag sediment outliers
}
\examples{
data("testData",package="sedReview")
x <- testData
sedOutlierFlags <- getSedOutliers(x)

}
