% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sedTS.R
\name{plot_sedTS}
\alias{plot_sedTS}
\title{plot_sedTS. Create timeseries plots for sediment parameters.}
\usage{
plot_sedTS(x, siteSelect = NULL, PDFout = NULL)
}
\arguments{
\item{x}{A \code{dataframe} output from \code{get_localNWIS}}

\item{siteSelect}{Character, site number to create plots for if \code{x} contains multiple sites. Default is \code{NULL}.}

\item{PDFout}{Character. File or full path name of file for plots. If \code{NULL}, the default, a list of the plots will be returned in R instead.}
}
\value{
If \code{PDFout = NULL}, list containing ggplot elements. If \code{PDFout} specified, a PDF document containing the plots.
}
\description{
Function to output timeseries plots for sediment parameters. Output is a list of plots or write to PDF.
}
\details{
Timeseries plots of SSC (P80154), sand/silt break (P70331), SSL (P80155), bedload (P80255) and bedload mass (P91145).
Not all parameters need to be present in \code{x} for plots to be created (eg. if only SSC present than only SSC timeseries plot returned)

If PDFout is not specified, than a list of the plots is returned. Plots (if applicable) are specified above. See example for more details.

Portions of code modified from \code{WQReview::qwtsPlot}.
}
\examples{
data("exampleData",package="sedReview")
x <- exampleData
sedTS <- plot_sedTS(exampleData, siteSelect = "05586300")
\dontrun{
# view plot in Rstudio
sedTS$SSC

# output to file on D drive
plot_sedTS(exampleData, siteSelect = "05586300", PDFout = "D:/timeseries.pdf")
}

}
