% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_boxcoef.R
\name{summary_boxcoef}
\alias{summary_boxcoef}
\title{summary_boxcoef}
\usage{
summary_boxcoef(x, timediff = 1, returnAllTables = FALSE)
}
\arguments{
\item{x}{A \code{dataframe} output from \code{get_localNWIS}}

\item{timediff}{Number of hours to look before and after a point sample for a paired cross-section sample.
Default is 1 (ie. look for a paired sample 1 hour before and 1 hour after a point sample timestamp)}

\item{returnAllTables}{Logical. If \code{TRUE}, return a list with the summary table and all individual site tables from the find_boxcoef function used internally.
Default is FALSE.}
}
\value{
A dataframe of the summary counts of box coefficient pairs, or a list of the summary and individual site box_coef data
}
\description{
Function to summarise box coefficient pairs for multiple sites and years utilizing the find_boxcoef function
}
\details{
Function summarises box coefficient pairs by each site and WY. See documentation for \code{find_boxcoef} function for more details.
}
\examples{
data('exampleData2', package = "sedReview")
x <- exampleData2
boxCoef <- summary_boxcoef(x)
boxCoef_list <- summary_boxcoef(x, returnAllTables = TRUE)
\dontrun{
View(boxCoef_list$summary)
}
}
\seealso{
\code{\link[sedReview]{get_localNWIS}}, \code{\link[sedReview]{find_boxcoef}}
}
